requirements.txt
django
djangorestframework
mysqlclient
django-background-tasks

cryptography

apscheduler
tweepy
pyyaml
python-dateutil
nagisa
